feat(FOROME-916): fix comments and reformat folder of main page
