+++ src/components/input-search/input-search.module.css
@@ -5,21 +5,13 @@
   font-size: 12px;
   line-height: 16px;
   border: 1px solid theme("colors.grey.blue");
-}
 
-.input-search:focus {
-  outline: none;
-}
+  &:focus {
+    outline: none;
+  }
 
-.input-search-big {
-  border-radius: 50vh;
-  width: 100%;
-  padding: 6px 35px 6px 12px;
-  font-size: 14px;
-  line-height: 20px;
-  border: 1px solid theme("colors.grey.blue");
-}
-
-.input-search-big:focus {
-  outline: none;
+  &_big {
+    font-size: 14px;
+    line-height: 20px;
+  }
 }
