+++ src/pages/main/components/sidebar/datasets-list/datasets-list.tsx
@@ -2,7 +2,7 @@ import { ReactElement } from 'react'
 import { observer } from 'mobx-react-lite'
 
 import dirinfoStore from '@store/dirinfo'
-import { datasetNameByKey } from '@pages/main/components/datasets-list/datasets-list.utils'
+import { datasetNameByKey } from '@pages/main/components/sidebar/datasets-list/datasets-list.utils'
 
 export const DatasetsList = observer((): ReactElement => {
   return (
